requirements.txt
pypdf==5.1.0
langchain==0.3.7
langchain-openai==0.2.8
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.19
